chore: refactor new list component
