+++ src/components/camel-list-page/CamelAppSummary.tsx
@@ -1,13 +1,7 @@
 import * as React from 'react';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  Flex,
-  FlexItem,
-  Label,
-  Split,
-  SplitItem,
-} from '@patternfly/react-core';
+import { Flex, FlexItem, Label, Split, SplitItem } from '@patternfly/react-core';
 import {
   GreenCheckCircleIcon,
   RedExclamationCircleIcon,
@@ -48,35 +42,51 @@ const CamelAppSummary: React.FC<CamelAppSummaryProps> = ({ data, onHealthFilter 
   }, [data]);
 
   return (
-    <Flex
-      spaceItems={{ default: 'spaceItemsLg' }}
-      alignItems={{ default: 'alignItemsCenter' }}
-    >
-
+    <Flex spaceItems={{ default: 'spaceItemsLg' }} alignItems={{ default: 'alignItemsCenter' }}>
       <FlexItem>
         <Split hasGutter>
           <SplitItem>
-            <Label color="green" icon={<GreenCheckCircleIcon />} isCompact onClick={() => onHealthFilter?.(HealthStatus.HEALTHY)}>
+            <Label
+              color="green"
+              icon={<GreenCheckCircleIcon />}
+              isCompact
+              onClick={() => onHealthFilter?.(HealthStatus.HEALTHY)}
+            >
               {healthCounts.healthy} {t('Healthy')}
             </Label>
           </SplitItem>
           {healthCounts.degraded > 0 && (
             <SplitItem>
-              <Label color="orange" icon={<YellowExclamationTriangleIcon />} isCompact onClick={() => onHealthFilter?.(HealthStatus.DEGRADED)}>
+              <Label
+                color="orange"
+                icon={<YellowExclamationTriangleIcon />}
+                isCompact
+                onClick={() => onHealthFilter?.(HealthStatus.DEGRADED)}
+              >
                 {healthCounts.degraded} {t('Degraded')}
               </Label>
             </SplitItem>
           )}
           {healthCounts.critical > 0 && (
             <SplitItem>
-              <Label color="red" icon={<RedExclamationCircleIcon />} isCompact onClick={() => onHealthFilter?.(HealthStatus.CRITICAL)}>
+              <Label
+                color="red"
+                icon={<RedExclamationCircleIcon />}
+                isCompact
+                onClick={() => onHealthFilter?.(HealthStatus.CRITICAL)}
+              >
                 {healthCounts.critical} {t('Critical')}
               </Label>
             </SplitItem>
           )}
           {healthCounts.unknown > 0 && (
             <SplitItem>
-              <Label color="grey" icon={<UnknownIcon />} isCompact onClick={() => onHealthFilter?.(HealthStatus.UNKNOWN)}>
+              <Label
+                color="grey"
+                icon={<UnknownIcon />}
+                isCompact
+                onClick={() => onHealthFilter?.(HealthStatus.UNKNOWN)}
+              >
                 {healthCounts.unknown} {t('Unknown')}
               </Label>
             </SplitItem>
